20250812
0812업데이트 완료...

+++ Player.h
@@ -15,5 +15,5 @@ public:
 	void White();
 	void Blue();
 
-	float Level;
+	int Level;
 };
